+++ app/src/main/java/eu/kanade/presentation/more/settings/screen/SettingsLibraryScreen.kt
@@ -206,10 +206,6 @@ object SettingsLibraryScreen : SearchableSettings {
                     title = stringResource(MR.strings.pref_library_update_show_tab_badge),
                 ),
                 Preference.PreferenceItem.SwitchPreference(
-                    preference = libraryPreferences.markDuplicateChapterRead(),
-                    title = stringResource(MR.strings.pref_mark_duplicate_chapter_read),
-                ),
-                Preference.PreferenceItem.SwitchPreference(
                     preference = libraryPreferences.showUpdatingProgressBanner(),
                     title = stringResource(MR.strings.pref_show_updating_progress_banner),
                 ),